server/requirements.txt
# Python 3.11 recommended
fastapi==0.110.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
scikit-learn==1.4.2
joblib==1.4.2
pillow==10.3.0
python-multipart==0.0.9
# Apple Silicon only (macOS); comment out if not on mac
tensorflow-macos==2.16.2
tensorflow-metal==1.2.0
